Исправлен баг с изображениями в списке. Исправлены ошибки при работе с FirestoreUtil после прошлых изменений.

+++ app/src/main/java/com/heads/thinking/headhelper/adapters/NewsRecyclerAdapter.kt
@@ -1,10 +1,13 @@
 package com.heads.thinking.headhelper.adapters
 
+import android.app.AlertDialog
 import android.content.Context
+import android.content.DialogInterface
 import android.content.Intent
 import android.media.Image
 import android.support.v7.widget.CardView
 import android.support.v7.widget.RecyclerView
+import android.text.Layout
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -30,19 +33,27 @@ class NewsRecyclerAdapter(val context: Context, var list:ArrayList<News>, val ne
 
     override fun onBindViewHolder(viewHolder: ViewHolder, position: Int) {
         viewHolder.bindItems(list[position])
-        viewHolder.deleteBtn.setOnClickListener {
-            val deletingNews : News = list[position]
-            FirestoreUtil.deleteNews(deletingNews.id, {
-                if(it) {
-                    Toast.makeText(App.instance, "Новость удалена", Toast.LENGTH_SHORT).show()
-                    if(list.size == 1)
-                        newsViewModel.updateListener()
-                    if(deletingNews.picturePath != null)
-                        StorageUtil.deleteNewsImage(deletingNews.picturePath!!, {})
-                }
-                else
-                    Toast.makeText(App.instance, "Не получилось удалить", Toast.LENGTH_SHORT).show()
-            })
+        viewHolder.newsCardView.setOnLongClickListener {
+            if (FirestoreUtil.currentUser?.privilege ?: 0 != 0) {
+                val builder: AlertDialog.Builder = AlertDialog.Builder(context)
+                builder.setTitle("Вы действительно хотите удалить новость?")
+                builder.setPositiveButton("Удалить", { dialogInterface: DialogInterface, i: Int ->
+                    val deletingNews: News = list[position]
+                    FirestoreUtil.deleteNews(deletingNews.id, {
+                        if (it) {
+                            Toast.makeText(App.instance, "Новость удалена", Toast.LENGTH_SHORT).show()
+                            if (list.size == 1)
+                                newsViewModel.updateListener()
+                            if (deletingNews.picturePath != null)
+                                StorageUtil.deleteNewsImage(deletingNews.picturePath!!, {})
+                        } else
+                            Toast.makeText(App.instance, "Не получилось удалить", Toast.LENGTH_SHORT).show()
+                    })
+                })
+                .setNegativeButton("Отмена", null)
+                builder.create().show()
+            }
+            true
         }
         viewHolder.newsCardView.setOnClickListener {
             context.startActivity(Intent(context, NewsViewerActivity::class.java).apply {
@@ -58,13 +69,11 @@ class NewsRecyclerAdapter(val context: Context, var list:ArrayList<News>, val ne
         lateinit var newsCardView: CardView
         lateinit var imageView: ImageView
         lateinit var itemHeader: TextView
-        lateinit var deleteBtn: ImageView
 
         fun bindItems(news: News) {
             imageView = itemView.findViewById(R.id.newsCircleImageView)
             itemHeader = itemView.findViewById(R.id.itemHeader)
             newsCardView = itemView.findViewById(R.id.newsCardView)
-            deleteBtn = itemView.findViewById(R.id.deleteNewsBtn)
 
             //load attribute
             itemHeader.text = news.tittle
@@ -72,6 +81,11 @@ class NewsRecyclerAdapter(val context: Context, var list:ArrayList<News>, val ne
                 GlideApp.with(imageView)
                       .load(StorageUtil.pathToReference(news.picturePath))
                       .into(imageView)
+            else {
+                GlideApp.with(imageView)
+                        .load(R.drawable.logo)
+                        .into(imageView)
+            }
         }
     }
 }
